8_FastAPI_Agent_Server/requirements.txt
# requirements.txt

python-dotenv
langchain
langchain-openai
langchain-chroma
pypdf
chromadb
uvicorn
fastapi